beginning of major reorganization
--converting to a divisional organization structure
--moving toward better projects manager
