+++ OpenAgua/conversions.py
@@ -1,5 +1,11 @@
-def get_coords(network):
-    coords = {}
-    for n in network['nodes']:
-        coords[n.id] = [float(n.x), float(n.y)] 
-    return coords
+from flask import session, url_for
+from functools import wraps
+
+def login_required(f):
+    @wraps(f)
+    def decorated_function(*args, **kwargs):
+        if 'logged_in' in session:
+            return f(*args, **kwargs)
+        else:
+            return redirect(url_for('index'))
+    return decorated_function
