+++ OpenAgua/model_dashboard/__init__.py
@@ -1,0 +1,9 @@
+from flask import Blueprint
+
+model_dashboard = Blueprint('model_dashboard',
+                       __name__,
+                       template_folder='templates',
+                       static_folder='static',
+                       static_url_path='/model_dashboard/static')
+
+from . import views
