+++ OpenAgua/network_editor/static/lib/leaflet/draw/images/spritesheet-2x.png
