+++ OpenAgua/network_editor/static/lib/leaflet/leaflet/images/marker-icon.png
