+++ OpenAgua/network_editor/templates/head_leaflet.html
@@ -1,0 +1,32 @@
+<!--Leaflet-->
+<link rel="stylesheet" href="{{ url_for('.static', filename='lib/leaflet/leaflet/leaflet.css') }}" />
+<script src="{{ url_for('.static', filename='lib/leaflet/leaflet/leaflet.js') }}"></script>
+
+<!--draw-->
+<link rel="stylesheet" href="{{ url_for('.static', filename='lib/leaflet/draw/leaflet.draw.css') }}" />
+<script src="{{ url_for('.static', filename='lib/leaflet/draw/leaflet.draw.js') }}"></script>
+
+<!--geometryutil-->
+<script src="{{ url_for('.static', filename='lib/leaflet/geometryutil/leaflet.geometryutil.js') }}"></script>
+
+<!--snap-->
+<script src="{{ url_for('.static', filename='lib/leaflet/snap/leaflet.snap.js') }}"></script>
+
+<!--fullscreen-->
+<link rel="stylesheet" href="{{ url_for('.static', filename='lib/leaflet/fullscreen/leaflet.fullscreen.css') }}" />
+<script src="{{ url_for('.static', filename='lib/leaflet/fullscreen/leaflet.fullscreen.min.js') }}"></script>
+
+<!--locate-->
+<link rel="stylesheet" href="{{ url_for('.static', filename='lib/leaflet/locate/L.Control.Locate.min.css') }}" />
+<script src="{{ url_for('.static', filename='lib/leaflet/locate/L.Control.Locate.min.js') }}"></script>
+
+<!--boxzoom-->
+<link rel="stylesheet" href="{{ url_for('.static', filename='lib/leaflet/boxzoom/leaflet-control-boxzoom.css') }}" />
+<script src="{{ url_for('.static', filename='lib/leaflet/boxzoom/leaflet-control-boxzoom.js') }}"></script>
+
+<!--contextmenu-->
+<link rel="stylesheet" href="{{ url_for('.static', filename='lib/leaflet/contextmenu/leaflet.contextmenu.min.css') }}" />
+<script src="{{ url_for('.static', filename='lib/leaflet/contextmenu/leaflet.contextmenu.min.js') }}"></script>
+
+<!--spin-->
+<script src="{{ url_for('.static', filename='lib/leaflet/spin/leaflet.spin.js') }}"></script>
