+++ OpenAgua/templates/navbar.html
@@ -12,7 +12,7 @@
     <div id="navbar" class="navbar-collapse collapse">
       <ul class="nav navbar-nav navbar-left">
       <li class="nav-item"><a class="nav-link" href="/home">Home</a></li>
-      <li class="nav-item"><a class="nav-link" href="/settings">Settings</a></li>
+      <li class="nav-item"><a class="nav-link" href="/projects">Projects</a></li>
       <li class="nav-item"><a class="nav-link" href="#">Help</a></li>
       </ul>  
       <ul class="nav navbar-nav navbar-right">
