+++ OpenAgua/templates/overview.html
@@ -1,7 +1,6 @@
 {% extends "master.html" %}
 {% block title %} - Network{% endblock %}
 {% block head_main %}
-{% include 'head_leaflet.html' %}
 {% endblock %}
 
 {% block content %}
